Update main.tf
main.tf

Here's the fun part. The part that initializes the server. It's also surprisingly short:

With all the work we've done in the other files, all we need to do here is describe the server we want.

Let's break down what this configuration is saying:

    We are defining an aws_instance with the unique Terraform identifier of my-test-instance

    That instance should use the AMI found in aws_ami.tf to initialize the server

    That instance should be a t2.micro (the cheapest AWS instance type)

    We've attached a Name tag to the instance, test-instance, for easy identification

This server won't do much - yet. As you'll notice, we haven't installed anything on it.

We could theoretically do that through a manual process, after we create the server. But that's not in the spirit of Infrastructure as Code!

We'll start adding things to this server in the next chapter. But before that happens, let's try out Terraform!